# docs/swagger.yaml
basePath: /
definitions:
  dto.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  pkg.PageInfo:
    properties:
      currentPage:
        type: integer
      nextPage:
        type: integer
      prevPage:
        type: integer
      totalData:
        type: integer
      totalPage:
        type: integer
    type: object
  pkg.Response:
    properties:
      error: {}
      message:
        type: string
      pageInfo:
        $ref: '#/definitions/pkg.PageInfo'
      result: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /users:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Get all users with pagination, sorting, and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 5
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: id
        description: Field to sort by
        in: query
        name: sortBy
        type: string
      - default: asc
        description: Order of sorting (asc or desc)
        in: query
        name: order
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - multipart/form-data
      description: Create a new user with the provided details
      parameters:
      - in: formData
        name: email
        type: string
      - in: formData
        name: first_name
        type: string
      - in: formData
        name: last_name
        type: string
      - in: formData
        name: password
        type: string
      - in: formData
        name: phone_number
        type: string
      - in: formData
        name: role
        type: string
      - description: Profile picture
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Creat new user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: Delete the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: get detail user
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get detail user
      tags:
      - Users
    patch:
      consumes:
      - multipart/form-data
      description: Update an existing user
      parameters:
      - in: formData
        name: email
        type: string
      - in: formData
        name: first_name
        type: string
      - in: formData
        name: last_name
        type: string
      - in: formData
        name: password
        type: string
      - in: formData
        name: phone_number
        type: string
      - description: Profile picture
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"